Added image test in demo.
Image is drawn and saved as png. It need slight adjustment in the area where it is drawn. Remember TODO checklist.

Everything has been converted into `int` instead of `double`. It's much easier to use `int` since they'll be used like pixels, anyway.

`GenerateCellList` added in `CellGenerator`. It needs to be generated using "Park-Miller Normal Distribution" (which I can't find, besides Normal Gaussian Distribution. That's the one that was linked in the instruction as well). Currently, it is generated randomly.

Documentation is required after `CellGenerator` class is completed. It should  also explain how  things works there.

+++ procedural_dungeon_generator/Common/Point.cs
@@ -5,10 +5,10 @@ using System.Text;
 namespace procedural_dungeon_generator.Common {
     public class Point {
 
-        public double X { get; }
-        public double Y { get; }
+        public int X { get; }
+        public int Y { get; }
 
-        public Point(double x, double y) {
+        public Point(int x, int y) {
             X = x;
             Y = y;
         }
